+++ packages/vm/CHANGELOG.md
@@ -13,7 +13,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 **CI and Test Improvements**
 
-- The `afterBlock` event has been changed from emitting an Object of shape `{ receipts, results }` to `RunBlockResult`, PR [#853](https://github.com/ethereumjs/ethereumjs-vm/pull/853)
+- `{ stateRoot, gasUsed, logsBloom, receiptRoot }` have been added to `RunBlockResult` and will be emitted with `afterBlock`, PR [#853](https://github.com/ethereumjs/ethereumjs-vm/pull/853)
 - Benchmark improvements and fixes, PR [#853](https://github.com/ethereumjs/ethereumjs-vm/pull/853)
 
 ### 5.1.0 - 2021-02-22